- fixed some problems with loading block compressed textures in dx12 - fixed bugs in mesh pipeline when importing/loading large meshes
Former-commit-id: 5ddf2df4224f1886b5ef98ec5a78e6d01298731e

+++ VEngine/src/Graphics/Mesh.h
@@ -47,7 +47,7 @@ namespace VEngine
 		glm::vec2 m_minTexCoord;
 		glm::vec2 m_maxTexCoord;
 		uint32_t m_vertexCount;
-		uint16_t m_indexCount;
+		uint32_t m_indexCount;
 		uint8_t *m_positions;
 		uint8_t *m_qtangents;
 		uint8_t *m_texCoords;
